+++ GyeongJu/food/food_data.py
@@ -1,7 +1,7 @@
 from bs4 import BeautifulSoup
 import time
 import requests
-from .models import *
+from .food_models import *
 
 user_agent = {"User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_4) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/83.0.4103.97 Safari/537.36"}
